+++ assets/BoxPage.fbb4752c.js
@@ -1,0 +1,20 @@
+var m=Object.defineProperty;var o=(e,a)=>m(e,"name",{value:a,configurable:!0});import{u as i,d as t,B as n}from"./index.cdfdb010.js";import{C as p}from"./ComponentInfo.d0daa232.js";import"./Link.7be7d999.js";import"./copyText.0c8ea5ff.js";import"./PaperCode.7a4736ea.js";function r(){const e=i();return t(n,{get sx(){return{width:300,height:300,backgroundColor:e.palette.primary.dark,"&:hover":{backgroundColor:e.palette.primary.main,opacity:.9}}}})}o(r,"BoxSx");r.code=`import Box from "@suid/material/Box";
+import useTheme from "@suid/material/styles/useTheme";
+
+export default function BoxSx() {
+  const theme = useTheme();
+  return (
+    <Box
+      sx={{
+        width: 300,
+        height: 300,
+        backgroundColor: theme.palette.primary.dark,
+        "&:hover": {
+          backgroundColor: theme.palette.primary.main,
+          opacity: 0.9,
+        },
+      }}
+    />
+  );
+}
+`;function d(){return t(p,{get name(){return n.name},docsApiName:"components/box",moreExamples:!1,examples:[{component:r,title:!1}]})}o(d,"BoxPage");export{d as default};