+++ assets/LinksPage.94dfda3c.js
@@ -1,0 +1,27 @@
+var a=Object.defineProperty;var r=(t,i)=>a(t,"name",{value:i,configurable:!0});import{L as e}from"./Link.7be7d999.js";import{C as l}from"./ComponentInfo.d0daa232.js";import{d as n,B as m}from"./index.cdfdb010.js";import"./copyText.0c8ea5ff.js";import"./PaperCode.7a4736ea.js";function o(){return n(m,{sx:{typography:"body1","& > :not(style) + :not(style)":{ml:2}},onClick:t=>t.preventDefault(),get children(){return[n(e,{href:"#",target:"none",children:"Link"}),n(e,{href:"#",color:"inherit",target:"none",children:'color="inherit"'}),n(e,{href:"#",variant:"body2",target:"none",children:'variant="body2"'})]}})}r(o,"BasicLinks");o.code=`import Box from "@suid/material/Box";
+import Link from "@suid/material/Link";
+
+export default function BasicLinks() {
+  return (
+    <Box
+      sx={{
+        typography: "body1",
+        "& > :not(style) + :not(style)": {
+          ml: 2,
+        },
+      }}
+      onClick={(event) => event.preventDefault()}
+    >
+      <Link href="#" target="none">
+        Link
+      </Link>
+      <Link href="#" color="inherit" target="none">
+        {'color="inherit"'}
+      </Link>
+      <Link href="#" variant="body2" target="none">
+        {'variant="body2"'}
+      </Link>
+    </Box>
+  );
+}
+`;function h(){return n(l,{get name(){return e.name},docsName:"links",examples:[o]})}r(h,"LinksPage");export{h as default};