+++ assets/TextFieldPage.e7887211.js
@@ -1,0 +1,232 @@
+var Te=Object.defineProperty;var s=(e,t)=>Te(e,"name",{value:t,configurable:!0});import{q as R,r as $,s as F,v as P,w as h,K as b,o as _,f as K,a as ue,m as p,af as Re,n as y,d as r,A as W,az as G,a8 as le,aA as $e,am as Le,aB as ce,ae as ke,t as N,ai as J,S as pe,ab as ze,a6 as m,aC as Be,B as Q}from"./index.cdfdb010.js";import{u as O,F as We}from"./useFormControl.9d480ed7.js";import{f as S,F as Pe}from"./formControlState.c1f71641.js";import{u as Oe}from"./useControlled.ae82aafa.js";import{C as Se}from"./ComponentInfo.d0daa232.js";import"./Link.7be7d999.js";import"./copyText.0c8ea5ff.js";import"./PaperCode.7a4736ea.js";function Ae(e){const t=e.element;let l="",n=0;t.addEventListener("beforeinput",()=>{l=t.value,n=t.selectionStart??l.length}),t.addEventListener("input",o=>{const a=t.value,i=t.selectionStart??a.length;t.value=l,t.selectionStart=t.selectionEnd=n,e.onChange(o,a,i)})}s(Ae,"prepareControlledInput");function Me(e){const[t,l]=e.valueAndStart,n=l??t.length;e.element.value=t,e.element.setSelectionRange(n,n)}s(Me,"setControlledInputValue");function qe(e){return $("MuiInputBase",e)}s(qe,"getInputBaseUtilityClass");const j=R("MuiInputBase",["root","formControl","focused","disabled","adornedStart","adornedEnd","error","sizeSmall","multiline","colorSecondary","fullWidth","hiddenLabel","input","inputSizeSmall","inputMultiline","inputTypeSearch","inputAdornedStart","inputAdornedEnd","inputHiddenLabel"]);function oe(e){return e!=null&&!(Array.isArray(e)&&e.length===0)}s(oe,"hasValue");function Ne(e,t=!1){return e&&(oe(e.value)&&e.value!==""||t&&oe(e.defaultValue)&&e.defaultValue!=="")}s(Ne,"isFilled");const re=F()({name:"MuiInputBase",propDefaults:({set:e})=>e({components:{},componentsProps:{},fullWidth:!1,inputComponent:"input",inputProps:{},multiline:!1,type:"text",disableInjectingGlobalStyles:!1}),selfPropNames:["aria-describedby","autoComplete","autoFocus","classes","color","components","componentsProps","defaultValue","disableInjectingGlobalStyles","disabled","endAdornment","error","fullWidth","id","inputComponent","inputProps","inputProps","inputRef","margin","maxRows","minRows","multiline","name","onBlur","onChange","onFocus","onKeyDown","onKeyUp","placeholder","readOnly","renderSuffix","required","rows","size","startAdornment","type","value"],utilityClass:qe,autoCallUseClasses:!1,slotClasses:e=>({root:["root",`color${P(e.color)}`,!!e.disabled&&"disabled",!!e.error&&"error",!!e.fullWidth&&"fullWidth",e.focused&&"focused",!!e.formControl&&"formControl",e.size==="small"&&"sizeSmall",e.multiline&&"multiline",!!e.startAdornment&&"adornedStart",!!e.endAdornment&&"adornedEnd",!!e.hiddenLabel&&"hiddenLabel"],input:["input",!!e.disabled&&"disabled",e.type==="search"&&"inputTypeSearch",e.multiline&&"inputMultiline",e.size==="small"&&"inputSizeSmall",!!e.hiddenLabel&&"inputHiddenLabel",!!e.startAdornment&&"inputAdornedStart",!!e.endAdornment&&"inputAdornedEnd"]})}),w=s((e,t)=>{const l=e.ownerState;return[t.root,!!l.formControl&&t.formControl,!!l.startAdornment&&t.adornedStart,!!l.endAdornment&&t.adornedEnd,!!l.error&&t.error,l.size==="small"&&t.sizeSmall,l.multiline&&t.multiline,l.color&&t[`color${P(l.color)}`],!!l.fullWidth&&t.fullWidth,!!l.hiddenLabel&&t.hiddenLabel]},"rootOverridesResolver"),V=s((e,t)=>{const l=e.ownerState;return[t.input,l.size==="small"&&t.inputSizeSmall,l.multiline&&t.inputMultiline,l.type==="search"&&t.inputTypeSearch,!!l.startAdornment&&t.inputAdornedStart,!!l.endAdornment&&t.inputAdornedEnd,!!l.hiddenLabel&&t.inputHiddenLabel]},"inputOverridesResolver"),H=h("div",{name:"MuiInputBase",slot:"Root",overridesResolver:w})(({theme:e,ownerState:t})=>({...e.typography.body1,color:e.palette.text.primary,lineHeight:"1.4375em",boxSizing:"border-box",position:"relative",cursor:"text",display:"inline-flex",alignItems:"center",[`&.${j.disabled}`]:{color:e.palette.text.disabled,cursor:"default"},...t.multiline&&{padding:"4px 0 5px",...t.size==="small"&&{paddingTop:1}},...t.fullWidth&&{width:"100%"}})),U=h("input",{name:"MuiInputBase",slot:"Input",overridesResolver:V})(({theme:e,ownerState:t})=>{const l=e.palette.mode==="light",n={color:"currentColor",opacity:l?.42:.5,transition:e.transitions.create("opacity",{duration:e.transitions.duration.shorter})},o={opacity:"0 !important"},a={opacity:l?.42:.5};return{font:"inherit",letterSpacing:"inherit",color:"currentColor",padding:"4px 0 5px",border:0,boxSizing:"content-box",background:"none",height:"1.4375em",margin:0,WebkitTapHighlightColor:"transparent",display:"block",minWidth:0,width:"100%",animationName:"mui-auto-fill-cancel",animationDuration:"10ms","&::-webkit-input-placeholder":n,"&::-moz-placeholder":n,"&:-ms-input-placeholder":n,"&::-ms-input-placeholder":n,"&:focus":{outline:0},"&:invalid":{boxShadow:"none"},"&::-webkit-search-decoration":{WebkitAppearance:"none"},[`label[data-shrink=false] + .${j.formControl} &`]:{"&::-webkit-input-placeholder":o,"&::-moz-placeholder":o,"&:-ms-input-placeholder":o,"&::-ms-input-placeholder":o,"&:focus::-webkit-input-placeholder":a,"&:focus::-moz-placeholder":a,"&:focus:-ms-input-placeholder":a,"&:focus::-ms-input-placeholder":a},[`&.${j.disabled}`]:{opacity:1,WebkitTextFillColor:e.palette.text.disabled},"&:-webkit-autofill":{animationDuration:"5000s",animationName:"mui-auto-fill"},...t.size==="small"&&{paddingTop:1},...t.multiline&&{height:"auto",resize:"none",padding:0,paddingTop:0},...t.type==="search"&&{MozAppearance:"textfield"}}}),we=s(()=>r($e,{styles:{"@keyframes mui-auto-fill":{from:{display:"block"}},"@keyframes mui-auto-fill-cancel":{from:{display:"block"}}}}),"inputGlobalStyles"),Y=re.component(s(function({allProps:t,otherProps:l,props:n}){const o=b(()=>n.inputProps.value!=null?n.inputProps.value:n.value),a=(o()??null)!==null,[i,f]=Oe({controlled:()=>o(),default:()=>n.defaultValue,name:"InputBase"}),c=Le({ref:d=>{typeof n.inputRef=="function"&&n.inputRef(d)}});let L;a?_(()=>{Ae({element:c.ref,onChange:(d,C,Ie)=>{L=Ie,typeof n.inputProps.onChange=="function"&&n.inputProps.onChange(d),f(C),typeof n.onChange=="function"&&n.onChange(d,C)}})}):_(()=>{c.ref.addEventListener("input",d=>{const C=c.ref.value;typeof n.inputProps.onChange=="function"&&n.inputProps.onChange(d),f(C),typeof n.onChange=="function"&&n.onChange(d,C)})}),K(d=>{if(a||d){const C=i();typeof C=="string"&&Me({element:c.ref,valueAndStart:[C,L]})}return!1},!0);const[I,z]=ue(!1),g=O(),x=S({props:t,muiFormControl:g,states:["color","disabled","error","hiddenLabel","size","required","filled"]}),v=p(()=>x,()=>({focused:g?g.focused:I()}));K(()=>{!g&&n.disabled&&I()&&(z(!1),typeof n.onBlur=="function"&&n.onBlur?.(null))});const ge=b(()=>g&&g.onFilled),he=b(()=>g&&g.onEmpty),A=s(d=>{Ne(d)?ge()?.():he()?.()},"checkDirty");Re(()=>{a&&A({value:i()})}),_(()=>{A(c.ref)});const Z=b(()=>n.multiline&&n.inputComponent==="input"),xe=b(()=>{const d=n.inputComponent;return Z(),d}),E=b(()=>{let d=n.inputProps;return Z()&&(n.rows?d={type:void 0,minRows:n.rows,maxRows:n.rows,...d}:d={type:void 0,maxRows:n.maxRows,minRows:n.minRows,...d}),p(d,()=>n.componentsProps.input||{})});K(()=>{g?.setAdornedStart(Boolean(n.startAdornment))});const M=p(t,()=>({...n,color:v.color||"primary",disabled:v.disabled,error:v.error,focused:v.focused,formControl:g,hiddenLabel:v.hiddenLabel,size:v.size})),ee=re.useClasses(M),te=b(()=>n.components.Root||H),D=b(()=>n.componentsProps.root||{}),ne=b(()=>n.components.Input||U),ve=p(()=>M,()=>D().ownerState||{}),Ce=p(()=>M,()=>E().ownerState||{}),ye=p(v,()=>({startAdornment:n.startAdornment})),Fe=b(()=>n.renderSuffix?.(ye));return[y((()=>{const d=y(()=>!n.disableInjectingGlobalStyles,!0);return()=>d()&&we()})()),r(G,p(D,l,{get component(){return te()}},()=>!le(te())&&{ownerState:ve},{onClick:d=>{c.ref&&d.currentTarget===d.target&&c.ref.focus(),typeof l.onClick=="function"&&l.onClick(d)},get className(){return W(ee.root,D().className,l.className)},get children(){return[y(()=>n.startAdornment),r(We.Provider,{value:void 0,get children(){return r(G,p({get component(){return ne()},ownerState:M,get["aria-invalid"](){return v.error},get["aria-describedby"](){return n["aria-describedby"]},get autocomplete(){return n.autoComplete},get autofocus(){return n.autoFocus},get disabled(){return v.disabled},get id(){return n.id},onAnimationStart:d=>{A(d.animationName==="mui-auto-fill-cancel"?c.ref:{value:"x"})},get name(){return n.name},get placeholder(){return n.placeholder},get readOnly(){return n.readOnly},get required(){return v.required},get rows(){return n.rows},get onKeyDown(){return n.onKeyDown},get onKeyUp(){return n.onKeyUp},get type(){return n.type}},E,()=>!le(ne())&&{as:xe(),ownerState:Ce},{ref:c,get className(){return W(ee.input,E().className)},onBlur:d=>{n.onBlur?.(d),typeof n.inputProps.onBlur=="function"&&n.inputProps.onBlur(d),g&&g.onBlur?g.onBlur():z(!1)},onInput:d=>{if(!a){const C=d.target||c.ref;if(C==null)throw new Error("MUI: Expected valid input target. Did you use a custom `inputComponent` and forget to forward refs? See https://mui.com/r/input-component-ref-interface for more info.");A({value:C.value})}},onFocus:d=>{if(v.disabled){d.stopPropagation();return}typeof n.onFocus=="function"&&n.onFocus(d),typeof n.inputProps.onFocus=="function"&&n.inputProps.onFocus(d),g&&g.onFocus?g.onFocus():z(!0)}}))}}),y(()=>n.endAdornment),y(()=>Fe())]}}))]},"InputBase"));function Ve(e){return $("MuiFilledInput",e)}s(Ve,"getFilledInputUtilityClass");const k=R("MuiFilledInput",["root","colorSecondary","underline","focused","disabled","adornedStart","adornedEnd","error","sizeSmall","multiline","hiddenLabel","input","inputSizeSmall","inputHiddenLabel","inputMultiline","inputAdornedStart","inputAdornedEnd"]),He=F()({name:"MuiFilledInput",propDefaults:({set:e})=>e({components:{},fullWidth:!1,inputComponent:"input",multiline:!1,type:"text",hiddenLabel:!1}),selfPropNames:["classes","disableUnderline","hiddenLabel"],utilityClass:Ve,slotClasses:e=>({root:["root",!e.disableUnderline&&"underline"],input:["input"]})}),Ue=h(H,{name:"MuiFilledInput",slot:"Root",overridesResolver:(e,t)=>{const{ownerState:l}=e;return[...w(e,t),!l.disableUnderline&&t.underline]}})(({theme:e,ownerState:t})=>{const l=e.palette.mode==="light",n=l?"rgba(0, 0, 0, 0.42)":"rgba(255, 255, 255, 0.7)",o=l?"rgba(0, 0, 0, 0.06)":"rgba(255, 255, 255, 0.09)";return{position:"relative",backgroundColor:o,borderTopLeftRadius:`${e.shape.borderRadius}px`,borderTopRightRadius:`${e.shape.borderRadius}px`,transition:e.transitions.create("background-color",{duration:e.transitions.duration.shorter,easing:e.transitions.easing.easeOut}),"&:hover":{backgroundColor:l?"rgba(0, 0, 0, 0.09)":"rgba(255, 255, 255, 0.13)","@media (hover: none)":{backgroundColor:o}},[`&.${k.focused}`]:{backgroundColor:o},[`&.${k.disabled}`]:{backgroundColor:l?"rgba(0, 0, 0, 0.12)":"rgba(255, 255, 255, 0.12)"},...!t.disableUnderline&&{"&:after":{borderBottom:`2px solid ${e.palette[t.color].main}`,left:0,bottom:0,content:'""',position:"absolute",right:0,transform:"scaleX(0)",transition:e.transitions.create("transform",{duration:e.transitions.duration.shorter,easing:e.transitions.easing.easeOut}),pointerEvents:"none"},[`&.${k.focused}:after`]:{transform:"scaleX(1)"},[`&.${k.error}:after`]:{borderBottomColor:e.palette.error.main,transform:"scaleX(1)"},"&:before":{borderBottom:`1px solid ${n}`,left:0,bottom:0,content:'"\\00a0"',position:"absolute",right:0,transition:e.transitions.create("border-bottom-color",{duration:e.transitions.duration.shorter}),pointerEvents:"none"},[`&:hover:not(.${k.disabled}):before`]:{borderBottom:`1px solid ${e.palette.text.primary}`},[`&.${k.disabled}:before`]:{borderBottomStyle:"dotted"}},...t.startAdornment&&{paddingLeft:12},...t.endAdornment&&{paddingRight:12},...t.multiline&&{padding:"25px 12px 8px",...t.size==="small"&&{paddingTop:21,paddingBottom:4},...t.hiddenLabel&&{paddingTop:16,paddingBottom:17}}}}),Ee=h(U,{name:"MuiFilledInput",slot:"Input",overridesResolver:V})(({theme:e,ownerState:t})=>({paddingTop:25,paddingRight:12,paddingBottom:8,paddingLeft:12,"&:-webkit-autofill":{WebkitBoxShadow:e.palette.mode==="light"?null:"0 0 0 100px #266798 inset",WebkitTextFillColor:e.palette.mode==="light"?null:"#fff",caretColor:e.palette.mode==="light"?null:"#fff",borderTopLeftRadius:"inherit",borderTopRightRadius:"inherit"},...t.size==="small"&&{paddingTop:21,paddingBottom:4},...t.hiddenLabel&&{paddingTop:16,paddingBottom:17},...t.multiline&&{paddingTop:0,paddingBottom:0,paddingLeft:0,paddingRight:0},...t.startAdornment&&{paddingLeft:0},...t.endAdornment&&{paddingRight:0},...t.hiddenLabel&&t.size==="small"&&{paddingTop:8,paddingBottom:9}})),De=He.component(s(function({allProps:t,classes:l,otherProps:n,props:o}){const a=b(()=>{const f={root:{ownerState:t},input:{ownerState:t}};return n.componentsProps?ce(n.componentsProps,f):f}),i=p(()=>o.classes||{},l);return r(Y,p(n,{get components(){return{Root:Ue,Input:Ee,...n.components||{}}},get componentsProps(){return a()},classes:i}))},"FilledInput"));function _e(e){return $("MuiFormHelperText",e)}s(_e,"getFormHelperTextUtilityClasses");const ie=R("MuiFormHelperText",["root","error","disabled","sizeSmall","sizeMedium","contained","focused","filled","required"]),Ke=N('<span class="notranslate">&#8203;</span>'),Xe=F()({name:"MuiFormHelperText",propDefaults:({set:e})=>e({component:"p"}),selfPropNames:["children","classes","disabled","error","filled","focused","margin","required","variant"],utilityClass:_e,slotClasses:e=>({root:["root",!!e.disabled&&"disabled",!!e.error&&"error",!!e.size&&`size${P(e.size)}`,e.contained&&"contained",!!e.focused&&"focused",!!e.filled&&"filled",!!e.required&&"required"]})}),Ge=h("p",{name:"MuiFormHelperText",slot:"Root",overridesResolver:(e,t)=>{const{ownerState:l}=e;return[t.root,l.size&&t[`size${P(l.size)}`],l.contained&&t.contained,l.filled&&t.filled]}})(({theme:e,ownerState:t})=>({color:e.palette.text.secondary,...e.typography.caption,textAlign:"left",marginTop:3,marginRight:0,marginBottom:0,marginLeft:0,[`&.${ie.disabled}`]:{color:e.palette.text.disabled},[`&.${ie.error}`]:{color:e.palette.error.main},...t.size==="small"&&{marginTop:4},...t.contained&&{marginLeft:14,marginRight:14}})),je=Xe.component(s(function({allProps:t,classes:l,otherProps:n,props:o}){const a=O(),i=S({props:t,muiFormControl:a,states:["variant","size","disabled","error","filled","focused","required"]}),f=p(t,()=>({contained:i.variant==="filled"||i.variant==="outlined",variant:i.variant,size:i.size,disabled:i.disabled,error:i.error,filled:i.filled,focused:i.focused,required:i.required})),c=ke(()=>o.children);return r(Ge,p(n,{get as(){return n.component},ownerState:f,get className(){return W(l.root,n.className)},get children(){return y(()=>c()===" ",!0)()?Ke.cloneNode(!0):c()}}))},"FormHelperText"));function Je(e){return $("MuiInput",e)}s(Je,"getInputUtilityClass");const q=R("MuiInput",["root","formControl","focused","disabled","colorSecondary","underline","error","sizeSmall","multiline","fullWidth","input","inputSizeSmall","inputMultiline","inputTypeSearch"]),Qe=F()({name:"MuiInput",propDefaults:({set:e})=>e({components:{},fullWidth:!1,inputComponent:"input",multiline:!1,type:"text"}),selfPropNames:["classes","disableUnderline"],utilityClass:Je,slotClasses:e=>({root:["root",!e.disableUnderline&&"underline"],input:["input"]})}),Ye=h(H,{skipProps:J.filter(e=>e!=="classes"),name:"MuiInput",slot:"Root",overridesResolver:(e,t)=>{const{ownerState:l}=e;return[...w(e,t),!l.disableUnderline&&t.underline]}})(({theme:e,ownerState:t})=>{const n=e.palette.mode==="light"?"rgba(0, 0, 0, 0.42)":"rgba(255, 255, 255, 0.7)";return{position:"relative",...t.formControl&&{"label + &":{marginTop:16}},...!t.disableUnderline&&{"&:after":{borderBottom:`2px solid ${e.palette[t.color].main}`,left:0,bottom:0,content:'""',position:"absolute",right:0,transform:"scaleX(0)",transition:e.transitions.create("transform",{duration:e.transitions.duration.shorter,easing:e.transitions.easing.easeOut}),pointerEvents:"none"},[`&.${q.focused}:after`]:{transform:"scaleX(1)"},[`&.${q.error}:after`]:{borderBottomColor:e.palette.error.main,transform:"scaleX(1)"},"&:before":{borderBottom:`1px solid ${n}`,left:0,bottom:0,content:'"\\00a0"',position:"absolute",right:0,transition:e.transitions.create("border-bottom-color",{duration:e.transitions.duration.shorter}),pointerEvents:"none"},[`&:hover:not(.${q.disabled}):before`]:{borderBottom:`2px solid ${e.palette.text.primary}`,"@media (hover: none)":{borderBottom:`1px solid ${n}`}},[`&.${q.disabled}:before`]:{borderBottomStyle:"dotted"}}}}),Ze=h(U,{name:"MuiInput",slot:"Input",overridesResolver:V})({}),et=Qe.component(s(function({classes:t,otherProps:l,props:n}){const o=b(()=>{const f={root:{ownerState:{disableUnderline:n.disableUnderline}}};return l.componentsProps?ce(l.componentsProps,f):f}),a=p(t,()=>n.classes||{});return r(Y,p(l,{get components(){return{Root:Ye,Input:Ze,...l.components||{}}},get componentsProps(){return o()},classes:a}))},"Input"));function tt(e){return $("MuiFormLabel",e)}s(tt,"getFormLabelUtilityClasses");const B=R("MuiFormLabel",["root","colorSecondary","focused","disabled","error","filled","required","asterisk"]),ae=F()({name:"MuiFormLabel",propDefaults:({set:e})=>e({component:"label"}),selfPropNames:["children","classes","color","disabled","error","filled","focused","required"],autoCallUseClasses:!1,utilityClass:tt,slotClasses:e=>({root:["root",`color${P(e.color)}`,!!e.disabled&&"disabled",!!e.error&&"error",!!e.filled&&"filled",!!e.focused&&"focused",!!e.required&&"required"],asterisk:["asterisk",!!e.error&&"error"]})}),nt=h("label",{name:"MuiFormLabel",slot:"Root",overridesResolver:({ownerState:e},t)=>({...t.root,...e.color==="secondary"&&t.colorSecondary,...e.filled&&t.filled})})(({theme:e,ownerState:t})=>({color:e.palette.text.secondary,...e.typography.body1,lineHeight:"1.4375em",padding:0,position:"relative",[`&.${B.focused}`]:{color:e.palette[t.color].main},[`&.${B.disabled}`]:{color:e.palette.text.disabled},[`&.${B.error}`]:{color:e.palette.error.main}})),lt=h("span",{name:"MuiFormLabel",slot:"Asterisk",overridesResolver:(e,t)=>t.asterisk})(({theme:e})=>({[`&.${B.error}`]:{color:e.palette.error.main}})),ot=ae.component(s(function({allProps:t,otherProps:l,props:n}){const o=O(),a=S({props:t,muiFormControl:o,states:["color","required","focused","disabled","error","filled"]}),i=p(t,()=>({color:a.color||"primary",disabled:a.disabled,error:a.error,filled:a.filled,focused:a.focused,required:a.required})),f=ae.useClasses(i);return r(nt,p(l,{get as(){return l.component},ownerState:i,get className(){return W(f.root,l.className)},get children(){return[y(()=>n.children),r(pe,{get when(){return a.required},get children(){return r(lt,{ownerState:i,"aria-hidden":!0,get className(){return f.asterisk},get children(){return["\u2009","*"]}})}})]}}))},"FormLabel"));function rt(e){return $("MuiInputLabel",e)}s(rt,"getInputLabelUtilityClasses");R("MuiInputLabel",["root","focused","disabled","error","required","asterisk","formControl","sizeSmall","shrink","animated","standard","filled","outlined"]);const de=F()({name:"MuiInputLabel",propDefaults:({set:e})=>e({disableAnimation:!1}),selfPropNames:["children","classes","color","disableAnimation","disabled","error","focused","margin","required","shrink","variant"],autoCallUseClasses:!1,utilityClass:rt,slotClasses:e=>({root:["root",!!e.formControl&&"formControl",!e.disableAnimation&&"animated",!!e.shrink&&"shrink",e.size==="small"&&"sizeSmall",!!e.variant&&e.variant],asterisk:[!!e.required&&"asterisk"]})}),it=h(ot,{skipProps:J.filter(e=>e!=="classes"),name:"MuiInputLabel",slot:"Root",overridesResolver:(e,t)=>{const{ownerState:l}=e;return[{[`& .${B.asterisk}`]:t.asterisk},t.root,l.formControl&&t.formControl,l.size==="small"&&t.sizeSmall,l.shrink&&t.shrink,!l.disableAnimation&&t.animated,t[l.variant]]}})(({theme:e,ownerState:t})=>({display:"block",transformOrigin:"top left",whiteSpace:"nowrap",overflow:"hidden",textOverflow:"ellipsis",maxWidth:"100%",...t.formControl&&{position:"absolute",left:0,top:0,transform:"translate(0, 20px) scale(1)"},...t.size==="small"&&{transform:"translate(0, 17px) scale(1)"},...t.shrink&&{transform:"translate(0, -1.5px) scale(0.75)",transformOrigin:"top left",maxWidth:"133%"},...!t.disableAnimation&&{transition:e.transitions.create(["color","transform","max-width"],{duration:e.transitions.duration.shorter,easing:e.transitions.easing.easeOut})},...t.variant==="filled"&&{zIndex:1,pointerEvents:"none",transform:"translate(12px, 16px) scale(1)",maxWidth:"calc(100% - 24px)",...t.size==="small"&&{transform:"translate(12px, 13px) scale(1)"},...t.shrink&&{userSelect:"none",pointerEvents:"auto",transform:"translate(12px, 7px) scale(0.75)",maxWidth:"calc(133% - 24px)",...t.size==="small"&&{transform:"translate(12px, 4px) scale(0.75)"}}},...t.variant==="outlined"&&{zIndex:1,pointerEvents:"none",transform:"translate(14px, 16px) scale(1)",maxWidth:"calc(100% - 24px)",...t.size==="small"&&{transform:"translate(14px, 9px) scale(1)"},...t.shrink&&{userSelect:"none",pointerEvents:"auto",maxWidth:"calc(133% - 24px)",transform:"translate(14px, -9px) scale(0.75)"}}})),at=de.component(s(function({allProps:t,props:l}){const n=O(),[,o]=ze(t,["disableAnimation","margin","shrink","variant"]),a=b(()=>{let I=l.shrink;return typeof I>"u"&&n&&(I=n.filled||n.focused||n.adornedStart),I}),i=S({props:t,muiFormControl:n,states:["size","variant","required"]}),f=p(t,()=>({formControl:n,shrink:a(),size:i.size,variant:i.variant,required:i.required})),c=de.useClasses(f),L=p(c,()=>l.classes||{});return r(it,p(o,{get["data-shrink"](){return a()},ownerState:f,classes:L}))},"InputLabel")),dt=N("<span></span>"),st=N('<span class="notranslate">&#8203;</span>'),ut=F()({name:"MuiNotchedOutline",selfPropNames:["disabled","error","focused","label","notched"]}),ct=h("fieldset")({textAlign:"left",position:"absolute",bottom:0,right:0,top:-5,left:0,margin:0,padding:"0 8px",pointerEvents:"none",borderRadius:"inherit",borderStyle:"solid",borderWidth:"1px",overflow:"hidden",minWidth:"0%"}),pt=h("legend")(({ownerState:e,theme:t})=>({float:"unset",...!e.withLabel&&{padding:0,lineHeight:"11px",transition:t.transitions.create("width",{duration:150,easing:t.transitions.easing.easeOut})},...e.withLabel&&{display:"block",width:"auto",padding:0,height:11,fontSize:"0.75em",visibility:"hidden",maxWidth:"0.01px",transition:t.transitions.create("max-width",{duration:50,easing:t.transitions.easing.easeOut}),whiteSpace:"nowrap","& > span":{paddingLeft:5,paddingRight:5,display:"inline-block"},...e.notched&&{maxWidth:"100%",transition:t.transitions.create("max-width",{duration:100,easing:t.transitions.easing.easeOut,delay:50})}}})),mt=ut.component(s(function({allProps:t,otherProps:l,props:n}){const o=b(()=>n.label!=null&&n.label!==""),a=p(t,()=>({withLabel:o()}));return r(ct,p(l,{"aria-hidden":!0,get className(){return l.className},ownerState:a,get children(){return r(pt,{ownerState:a,get children(){return r(pe,{get when(){return o()},get fallback(){return st.cloneNode(!0)},get children(){const i=dt.cloneNode(!0);return m(i,()=>n.label),i}})}})}}))},"NotchedOutline"));function ft(e){return $("MuiOutlinedInput",e)}s(ft,"getOutlinedInputUtilityClass");const T=R("MuiOutlinedInput",["root","colorSecondary","focused","disabled","adornedStart","adornedEnd","error","sizeSmall","multiline","notchedOutline","input","inputSizeSmall","inputMultiline","inputAdornedStart","inputAdornedEnd"]),bt=F()({name:"MuiOutlinedInput",propDefaults:({set:e})=>e({components:{},fullWidth:!1,inputComponent:"input",multiline:!1,type:"text"}),selfPropNames:["classes","label","notched"],utilityClass:ft,slotClasses:()=>({root:["root"],notchedOutline:["notchedOutline"],input:["input"]})}),gt=h(H,{skipProps:J.filter(e=>e!=="classes"),name:"MuiOutlinedInput",slot:"Root",overridesResolver:w})(({theme:e,ownerState:t})=>{const l=e.palette.mode==="light"?"rgba(0, 0, 0, 0.23)":"rgba(255, 255, 255, 0.23)";return{position:"relative",borderRadius:e.shape.borderRadius,[`&:hover .${T.notchedOutline}`]:{borderColor:e.palette.text.primary},"@media (hover: none)":{[`&:hover .${T.notchedOutline}`]:{borderColor:l}},[`&.${T.focused} .${T.notchedOutline}`]:{borderColor:e.palette[t.color].main,borderWidth:"2px"},[`&.${T.error} .${T.notchedOutline}`]:{borderColor:e.palette.error.main},[`&.${T.disabled} .${T.notchedOutline}`]:{borderColor:e.palette.action.disabled},...t.startAdornment&&{paddingLeft:14},...t.endAdornment&&{paddingRight:14},...t.multiline&&{padding:"16.5px 14px",...t.size==="small"&&{padding:"8.5px 14px"}}}}),ht=h(mt,{name:"MuiOutlinedInput",slot:"NotchedOutline",overridesResolver:(e,t)=>t.notchedOutline})(({theme:e})=>({borderColor:e.palette.mode==="light"?"rgba(0, 0, 0, 0.23)":"rgba(255, 255, 255, 0.23)"})),xt=h(U,{name:"MuiOutlinedInput",slot:"Input",overridesResolver:V})(({theme:e,ownerState:t})=>({padding:"16.5px 14px","&:-webkit-autofill":{WebkitBoxShadow:e.palette.mode==="light"?null:"0 0 0 100px #266798 inset",WebkitTextFillColor:e.palette.mode==="light"?null:"#fff",caretColor:e.palette.mode==="light"?null:"#fff",borderRadius:"inherit"},...t.size==="small"&&{padding:"8.5px 14px"},...t.multiline&&{padding:0},...t.startAdornment&&{paddingLeft:0},...t.endAdornment&&{paddingRight:0}})),vt=bt.component(s(function({allProps:t,classes:l,otherProps:n,props:o}){const a=O(),i=S({props:t,muiFormControl:a,states:["required"]}),f=b(()=>{const c=o.label;return c!=null&&c!==""&&i.required?[c,"\xA0","*"]:c});return r(Y,p({renderSuffix:c=>r(ht,{get className(){return l.notchedOutline},get label(){return f()},get notched(){return y(()=>typeof o.notched<"u",!0)()?o.notched:Boolean(c.startAdornment||c.filled||c.focused)}})},n,{get components(){return{Root:gt,Input:xt,...n.components||{}}},get classes(){return{...l,...o.classes,notchedOutline:null}}}))},"OutlinedInput"));function Ct(e){return $("MuiTextField",e)}s(Ct,"getTextFieldUtilityClass");R("MuiTextField",["root"]);function yt(e){return b(()=>e?.()??`mui-${Be()}`)}s(yt,"createUniqueId");const Ft=F()({name:"MuiTextField",propDefaults:({set:e})=>e({autoFocus:!1,color:"primary",disabled:!1,error:!1,fullWidth:!1,multiline:!1,required:!1,select:!1,variant:"outlined"}),selfPropNames:["FormHelperTextProps","InputLabelProps","InputProps","autoComplete","autoFocus","children","classes","color","defaultValue","disabled","error","fullWidth","helperText","id","inputProps","inputRef","label","maxRows","minRows","multiline","name","onBlur","onChange","onFocus","placeholder","required","rows","select","size","type","value","variant"],utilityClass:Ct,slotClasses:()=>({root:["root"]})}),It={standard:et,filled:De,outlined:vt},Tt=h(Pe,{name:"MuiTextField",slot:"Root",overridesResolver:(e,t)=>t.root})({}),u=Ft.component(s(function({allProps:t,classes:l,otherProps:n,props:o}){const a=b(()=>{const x={};return o.variant==="outlined"&&(o.InputLabelProps&&typeof o.InputLabelProps.shrink<"u"&&(x.notched=o.InputLabelProps.shrink),x.label=o.label),x}),i=yt(()=>o.id),f=b(()=>o.helperText&&i()?`${i()}-helper-text`:void 0),c=b(()=>o.label&&i()?`${i()}-label`:void 0),L=b(()=>It[o.variant]),I=s(()=>r(G,p({get component(){return L()},get["aria-describedby"](){return f()},get autoComplete(){return o.autoComplete},get autoFocus(){return o.autoFocus},get defaultValue(){return o.defaultValue},get fullWidth(){return o.fullWidth},get multiline(){return o.multiline},get name(){return o.name},get rows(){return o.rows},get maxRows(){return o.maxRows},get minRows(){return o.minRows},get type(){return o.type},get value(){return o.value},get id(){return i()},get inputRef(){return o.inputRef},get onBlur(){return o.onBlur},get onChange(){return o.onChange},get onFocus(){return o.onFocus},get placeholder(){return o.placeholder},get inputProps(){return o.inputProps}},a,()=>o.InputProps||{})),"InputElement"),z=s(()=>{const x=o.label;return x!=null&&x!==""&&r(at,p({get htmlFor(){return i()},get id(){return c()}},()=>o.InputLabelProps||{},{children:x}))},"label"),g=s(()=>{const x=o.helperText;if(x)return r(je,p({get id(){return f()}},()=>o.FormHelperTextProps||{},{children:x}))},"helperText");return r(Tt,p({get className(){return W(l.root,n.className)},get disabled(){return o.disabled},get error(){return o.error},get fullWidth(){return o.fullWidth},get required(){return o.required},get color(){return o.color},get variant(){return o.variant},ownerState:t},n,{get children(){return[y(()=>z()),y(()=>I()),y(()=>g())]}}))},"TextField"));function me(){return r(Q,{component:"form",sx:{"& > :not(style)":{m:1,width:"25ch"}},noValidate:!0,autocomplete:"off",get children(){return[r(u,{id:"outlined-basic",label:"Outlined",variant:"outlined",fullWidth:!0}),r(u,{id:"filled-basic",label:"Filled",variant:"filled"}),r(u,{id:"standard-basic",label:"Standard",variant:"standard"})]}})}s(me,"BasicTextFields");me.code=`import Box from "@suid/material/Box";
+import TextField from "@suid/material/TextField";
+
+export default function BasicTextFields() {
+  return (
+    <Box
+      component="form"
+      sx={{
+        "& > :not(style)": { m: 1, width: "25ch" },
+      }}
+      noValidate
+      autocomplete="off"
+    >
+      <TextField
+        id="outlined-basic"
+        label="Outlined"
+        variant="outlined"
+        fullWidth
+      />
+      <TextField id="filled-basic" label="Filled" variant="filled" />
+      <TextField id="standard-basic" label="Standard" variant="standard" />
+    </Box>
+  );
+}
+`;const X=N("<div></div>");function fe(){return r(Q,{component:"form",sx:{maxWidth:764,[`& .${u}`]:{m:1,width:"25ch"}},noValidate:!0,autocomplete:"off",get children(){return[(()=>{const e=X.cloneNode(!0);return m(e,r(u,{required:!0,id:"outlined-required",label:"Required",defaultValue:"Hello World"}),null),m(e,r(u,{disabled:!0,id:"outlined-disabled",label:"Disabled",defaultValue:"Hello World"}),null),m(e,r(u,{id:"outlined-password-input",label:"Password",type:"password",autoComplete:"current-password"}),null),m(e,r(u,{id:"outlined-read-only-input",label:"Read Only",defaultValue:"Hello World",InputProps:{readOnly:!0}}),null),m(e,r(u,{id:"outlined-number",label:"Number",type:"number",InputLabelProps:{shrink:!0}}),null),m(e,r(u,{id:"outlined-search",label:"Search field",type:"search"}),null),m(e,r(u,{id:"outlined-helperText",label:"Helper text",defaultValue:"Default Value",helperText:"Some important text"}),null),e})(),(()=>{const e=X.cloneNode(!0);return m(e,r(u,{required:!0,id:"filled-required",label:"Required",defaultValue:"Hello World",variant:"filled"}),null),m(e,r(u,{disabled:!0,id:"filled-disabled",label:"Disabled",defaultValue:"Hello World",variant:"filled"}),null),m(e,r(u,{id:"filled-password-input",label:"Password",type:"password",autoComplete:"current-password",variant:"filled"}),null),m(e,r(u,{id:"filled-read-only-input",label:"Read Only",defaultValue:"Hello World",InputProps:{readOnly:!0},variant:"filled"}),null),m(e,r(u,{id:"filled-number",label:"Number",type:"number",InputLabelProps:{shrink:!0},variant:"filled"}),null),m(e,r(u,{id:"filled-search",label:"Search field",type:"search",variant:"filled"}),null),m(e,r(u,{id:"filled-helperText",label:"Helper text",defaultValue:"Default Value",helperText:"Some important text",variant:"filled"}),null),e})(),(()=>{const e=X.cloneNode(!0);return m(e,r(u,{required:!0,id:"standard-required",label:"Required",defaultValue:"Hello World",variant:"standard"}),null),m(e,r(u,{disabled:!0,id:"standard-disabled",label:"Disabled",defaultValue:"Hello World",variant:"standard"}),null),m(e,r(u,{id:"standard-password-input",label:"Password",type:"password",autoComplete:"current-password",variant:"standard"}),null),m(e,r(u,{id:"standard-read-only-input",label:"Read Only",defaultValue:"Hello World",InputProps:{readOnly:!0},variant:"standard"}),null),m(e,r(u,{id:"standard-number",label:"Number",type:"number",InputLabelProps:{shrink:!0},variant:"standard"}),null),m(e,r(u,{id:"standard-search",label:"Search field",type:"search",variant:"standard"}),null),m(e,r(u,{id:"standard-helperText",label:"Helper text",defaultValue:"Default Value",helperText:"Some important text",variant:"standard"}),null),e})()]}})}s(fe,"FormPropsTextFields");fe.code=`import Box from "@suid/material/Box";
+import TextField from "@suid/material/TextField";
+
+export default function FormPropsTextFields() {
+  return (
+    <Box
+      component="form"
+      sx={{
+        maxWidth: 764,
+        [\`& .\${TextField}\`]: { m: 1, width: "25ch" },
+      }}
+      noValidate
+      autocomplete="off"
+    >
+      <div>
+        <TextField
+          required
+          id="outlined-required"
+          label="Required"
+          defaultValue="Hello World"
+        />
+        <TextField
+          disabled
+          id="outlined-disabled"
+          label="Disabled"
+          defaultValue="Hello World"
+        />
+        <TextField
+          id="outlined-password-input"
+          label="Password"
+          type="password"
+          autoComplete="current-password"
+        />
+        <TextField
+          id="outlined-read-only-input"
+          label="Read Only"
+          defaultValue="Hello World"
+          InputProps={{
+            readOnly: true,
+          }}
+        />
+        <TextField
+          id="outlined-number"
+          label="Number"
+          type="number"
+          InputLabelProps={{
+            shrink: true,
+          }}
+        />
+        <TextField id="outlined-search" label="Search field" type="search" />
+        <TextField
+          id="outlined-helperText"
+          label="Helper text"
+          defaultValue="Default Value"
+          helperText="Some important text"
+        />
+      </div>
+      <div>
+        <TextField
+          required
+          id="filled-required"
+          label="Required"
+          defaultValue="Hello World"
+          variant="filled"
+        />
+        <TextField
+          disabled
+          id="filled-disabled"
+          label="Disabled"
+          defaultValue="Hello World"
+          variant="filled"
+        />
+        <TextField
+          id="filled-password-input"
+          label="Password"
+          type="password"
+          autoComplete="current-password"
+          variant="filled"
+        />
+        <TextField
+          id="filled-read-only-input"
+          label="Read Only"
+          defaultValue="Hello World"
+          InputProps={{
+            readOnly: true,
+          }}
+          variant="filled"
+        />
+        <TextField
+          id="filled-number"
+          label="Number"
+          type="number"
+          InputLabelProps={{
+            shrink: true,
+          }}
+          variant="filled"
+        />
+        <TextField
+          id="filled-search"
+          label="Search field"
+          type="search"
+          variant="filled"
+        />
+        <TextField
+          id="filled-helperText"
+          label="Helper text"
+          defaultValue="Default Value"
+          helperText="Some important text"
+          variant="filled"
+        />
+      </div>
+      <div>
+        <TextField
+          required
+          id="standard-required"
+          label="Required"
+          defaultValue="Hello World"
+          variant="standard"
+        />
+        <TextField
+          disabled
+          id="standard-disabled"
+          label="Disabled"
+          defaultValue="Hello World"
+          variant="standard"
+        />
+        <TextField
+          id="standard-password-input"
+          label="Password"
+          type="password"
+          autoComplete="current-password"
+          variant="standard"
+        />
+        <TextField
+          id="standard-read-only-input"
+          label="Read Only"
+          defaultValue="Hello World"
+          InputProps={{
+            readOnly: true,
+          }}
+          variant="standard"
+        />
+        <TextField
+          id="standard-number"
+          label="Number"
+          type="number"
+          InputLabelProps={{
+            shrink: true,
+          }}
+          variant="standard"
+        />
+        <TextField
+          id="standard-search"
+          label="Search field"
+          type="search"
+          variant="standard"
+        />
+        <TextField
+          id="standard-helperText"
+          label="Helper text"
+          defaultValue="Default Value"
+          helperText="Some important text"
+          variant="standard"
+        />
+      </div>
+    </Box>
+  );
+}
+`;function se(e){return e.split(" ").map(t=>t.slice(0,1).toUpperCase()+t.slice(1).toLowerCase()).join(" ")}s(se,"capitalizeWords");function be(){const[e,t]=ue(se("Cat in the Hat"));return r(Q,{component:"form",sx:{"& > :not(style)":{m:1,width:"25ch"}},novalidate:!0,autocomplete:"off",get children(){return[r(u,{id:"outlined-name",label:"Name",get value(){return e()},onChange:(l,n)=>{t(se(n))}}),r(u,{id:"outlined-uncontrolled",label:"Uncontrolled",defaultValue:"foo"})]}})}s(be,"StateTextFields");be.code=`import Box from "@suid/material/Box";
+import TextField from "@suid/material/TextField";
+import { createSignal } from "solid-js";
+
+function capitalizeWords(value: string) {
+  return value
+    .split(" ")
+    .map((v) => v.slice(0, 1).toUpperCase() + v.slice(1).toLowerCase())
+    .join(" ");
+}
+
+export default function StateTextFields() {
+  const [name, setName] = createSignal(capitalizeWords("Cat in the Hat"));
+
+  return (
+    <Box
+      component="form"
+      sx={{
+        "& > :not(style)": { m: 1, width: "25ch" },
+      }}
+      novalidate
+      autocomplete="off"
+    >
+      <TextField
+        id="outlined-name"
+        label="Name"
+        value={name()}
+        onChange={(event, value) => {
+          setName(capitalizeWords(value));
+        }}
+      />
+      <TextField
+        id="outlined-uncontrolled"
+        label="Uncontrolled"
+        defaultValue="foo"
+      />
+    </Box>
+  );
+}
+`;function St(){return r(Se,{get name(){return u.name},docsName:"text-fields",examples:[me,{title:"Form props",component:fe},{title:"Uncontrolled vs. Controlled",component:be}]})}s(St,"TextFieldPage");export{St as default};