+++ assets/UsagePage.799f7437.js
@@ -1,0 +1,12 @@
+var a=Object.defineProperty;var r=(e,o)=>a(e,"name",{value:o,configurable:!0});import{L as i}from"./Link.7be7d999.js";import{d as t,T as n,J as s}from"./index.cdfdb010.js";import{P as l}from"./PaperCode.7a4736ea.js";var m=`// @refresh
+import Button from "@suid/material/Button";
+import { render } from "solid-js/web";
+
+function App() {
+  return <Button variant="contained">Hello World</Button>;
+}
+
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+render(() => <App />, document.getElementById("root")!);
+`;function g(){return[t(n,{component:"h1",variant:"h4",sx:{mt:1},children:"Usage"}),t(n,{variant:"body1",sx:{mt:2},get children(){return["For a completed usage guide, please go to"," ",t(i,{href:"https://mui.com",target:"_blank",children:"MUI"}),"."]}}),t(n,{component:"h2",variant:"h5",sx:{mt:2,mb:1},children:"Quick usage"}),t(l,{language:"tsx",get value(){return m.split(/\r?\n/).filter(e=>!e.trim().startsWith("// eslint-disable")).join(`
+`)}}),t(s,{sx:{mt:2}})]}r(g,"UsagePage");export{g as default};